refactor(shuttle): refactor to shuttlefee hook
